adding a lot of attributes to property object

+++ database/migrations/2019_12_29_112113_create_listings_table.php
@@ -13,24 +13,33 @@ class CreateListingsTable extends Migration
      */
     public function up()
     {
-        Schema::create('listings', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->string('name');
-            $table->string('phone');
-            $table->string('mobile_phone');
-            $table->text('address');
-            $table->string('city');
-            $table->string('state');
-            $table->string('country');
-            $table->unsignedBigInteger('party_id')->nullable();
-            $table->unsignedBigInteger('user_id');
-            $table->timestamps();
+        if(!Schema::hasTable('listings')) {
+            Schema::create('listings', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->string('purpose');
+                $table->string('property_type');
+                $table->string('property_type_name');
+                $table->text('location');
+                $table->string('city');
+                $table->string('property_title');
+                $table->text('property_description')->nullable();
+                $table->string('property_price');
+                $table->string('land_area');
+                $table->string('area_unit');
+                $table->string('bedrooms');
+                $table->string('bathrooms');
+                $table->string('expires_after');
+                $table->string('images')->nullable();
+                $table->unsignedBigInteger('party_id')->nullable();
+                $table->unsignedBigInteger('user_id');
+                $table->timestamps();
 
-            // primary key foreign key relations
+                // primary key foreign key relations
 
-            $table->foreign('party_id')->references('id')->on('parties')->onDelete('cascade');
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-        });
+                $table->foreign('party_id')->references('id')->on('parties')->onDelete('cascade');
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+            });
+        }
     }
 
     /**
